use int16 for boostedTau category

+++ PhysicsTools/NanoAOD/python/boostedTaus_cff.py
@@ -55,7 +55,7 @@ _boostedTauVarsMVAIsoDr03 = cms.PSet(
 #AntiEle MVA 2018 variables
 _boostedTauVarsAntiEleMVA = cms.PSet(
        rawAntiEle2018 = Var("tauID('againstElectronMVA6Raw')", float, doc= "Anti-electron MVA discriminator V6 raw output discriminator (2018)", precision=10),
-       rawAntiEleCat2018 = Var("tauID('againstElectronMVA6category')", "int8", doc="Anti-electron MVA discriminator V6 category (2018)"),
+       rawAntiEleCat2018 = Var("tauID('againstElectronMVA6category')", "int16", doc="Anti-electron MVA discriminator V6 category (2018)"),
        idAntiEle2018 = _tauIdWPMask("againstElectron%sMVA6", choices=("VLoose","Loose","Medium","Tight","VTight"), doc= "Anti-electron MVA discriminator V6 (2018)")
 )
 
@@ -80,7 +80,7 @@ run2_nanoAOD_106Xv2.toModify(
     idMVAnewDM2017v2 = _tauIdWPMask("by%sIsolationMVArun2017v2DBnewDMwLT2017", choices=("VVLoose","VLoose","Loose","Medium","Tight","VTight","VVTight"),doc="IsolationMVArun2DBnewDMwLT ID working point (2017v2)"),
     idMVAoldDM2017v2 = _tauIdWPMask("by%sIsolationMVArun2017v2DBoldDMwLT2017",choices=("VVLoose","VLoose","Loose","Medium","Tight","VTight","VVTight"),doc="IsolationMVArun2DBoldDMwLT ID working point (2017v2)"),
     rawAntiEle2018 = Var("tauID('againstElectronMVA6Raw2018')", float, doc= "Anti-electron MVA discriminator V6 raw output discriminator (2018)", precision=10),
-    rawAntiEleCat2018 = Var("tauID('againstElectronMVA6category2018')", "int8", doc="Anti-electron MVA discriminator V6 category (2018)"),
+    rawAntiEleCat2018 = Var("tauID('againstElectronMVA6category2018')", "int16", doc="Anti-electron MVA discriminator V6 category (2018)"),
     idAntiEle2018 = _tauIdWPMask("againstElectron%sMVA62018", choices=("VLoose","Loose","Medium","Tight","VTight"), doc= "Anti-electron MVA discriminator V6 (2018)")
 )
